refactor(counting section): refactor counting section components structure
