+++ src/components/organisms/counting-section/list-counting-section.tsx
@@ -1,4 +1,6 @@
-import { type FC, type ReactElement, useEffect, useState } from 'react'
+import { type FC, useEffect, useState } from 'react'
+
+import ItemCountingSection from '@components/organisms/counting-section/item-counting-section'
 
 import { type ListCalculations } from '@libs/supabase/database.types'
 import { supabase } from '@libs/supabase/supabase'
@@ -20,21 +22,9 @@ const ListCountingSection: FC = () => {
     void getData()
   }, [])
 
-  // TODO: Move this to a new component
-  const renderListItem = (item: ListCalculations): ReactElement => (
-    <div key={item.id}>
-      <p>{item.id}</p>
-      <p>{item.country}</p>
-      <p>{item.arrival}</p>
-      <p>{item.departure}</p>
-      <p>{item.created_at}</p>
-      <p>{item.updated_at}</p>
-    </div>
-  )
-
   return (
     <>
-      {list?.map(renderListItem)}
+      {list?.map((item) => <ItemCountingSection key={item.id} item={item} />)}
     </>
   )
 }
