+++ src/pages/index.astro
@@ -4,7 +4,6 @@ import { t } from 'i18next'
 import Hero from '@molecules/hero.astro'
 
 import CountingSection from '@organisms/sections/counting-section'
-import ListCountingSection from '@organisms/sections/list-counting-section'
 
 import MainLayout from '@layouts/main-layout.astro'
 ---
@@ -12,12 +11,5 @@ import MainLayout from '@layouts/main-layout.astro'
 <MainLayout title={t('home:title')}>
   <Hero />
   <!-- TODO: pendding sections component improve order -->
-  <section
-    id="calc"
-    class="
-      container mx-auto flex min-h-screen flex-col flex-wrap items-center justify-center py-[140px] sm:py-[100px]"
-  >
-    <CountingSection client:idle t={t('home:counting', { returnObjects: true })} />
-    <ListCountingSection client:idle />
-  </section>
+  <CountingSection client:idle t={t('home:counting', { returnObjects: true })} />
 </MainLayout>
